Add files via upload
CookAloud is a mobile app developed for Android devices. The main idea behind the app is that it enables you to follow recipe's instructions without having physical contact with your phone. Instead of touching your phone everytime you want to switch to the next instruction all you need to do is just speak. The app provides users with a set of voice commands which they can use to control the app. For example, by saying "next" the app will move on to the next instruction in the recipe and tell you out loud what you need to do now. The app is really easy to use and it makes the cooking process very smooth.